requirements.txt
fastapi
uvicorn
aiogram
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
pydantic-settings
python-dotenv
dependency-injector
